# Conflicts:
#	pom.xml
#	src/main/java/io/khasang/moika/config/security/SecurityConfig.java
#	src/main/java/io/khasang/moika/controller/UserController.java
#	src/test/java/io/khasang/moika/service/impl/UserIntegrationTest.java

+++ src/main/java/io/khasang/moika/service/CompanyService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface CompanyService {
     void addCompany(Company company);
     void addCompany(Company company, String name);
-    Company getCompanyById(int id);
+    Company getCompanyById(long id);
     void updateCompany(Company company);
-    void deleteCompany(int id);
+    void deleteCompany(long id);
     List<Company> getCompanyGazpromList();
     void addButterfly(Butterfly butterfly);
 }
